requirements.txt
Flask==2.0.2
FLASK-SQLAlchemy==2.5.1
flask-restx==0.5.1
psycopg2-binary==2.8.6
python-dotenv
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-bcrypt==0.7.1
gunicorn==20.1.0
Werkzeug~=2.0.0
kafka-python==2.0.2
redis==4.1.4
pyjwt==2.3.0
